requirements.txt
opencv-python
ultralytics
supervision
requests
onvif-zeep